renamed local requests module
